LPS-206234 Move common logic to base
