+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/notifications/Notifications.js
@@ -3,16 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {useResource} from '@clayui/data-provider';
 import React, {useContext, useEffect, useState} from 'react';
 
 import {DefinitionBuilderContext} from '../../../../../DefinitionBuilderContext';
-import {contextUrl} from '../../../../../constants';
-import {
-	headers,
-	retrieveAccountRoles,
-	userBaseURL,
-} from '../../../../../util/fetchUtil';
 import {DiagramBuilderContext} from '../../../../DiagramBuilderContext';
 import NotificationsInfo from './NotificationsInfo';
 
@@ -21,42 +14,9 @@ const Notifications = (props) => {
 	const {selectedItem} = useContext(DiagramBuilderContext);
 
 	const {notifications} = selectedItem?.data;
-
-	const [accountRoles, setAccountRoles] = useState([]);
-	const [networkStatus, setNetworkStatus] = useState(4);
 	const [sections, setSections] = useState([{identifier: `${Date.now()}-0`}]);
 
-	const {resource} = useResource({
-		fetchOptions: {
-			headers: {
-				...headers,
-				'accept': `application/json`,
-				'x-csrf-token': Liferay.authToken,
-			},
-		},
-		fetchPolicy: 'cache-first',
-		link: `${window.location.origin}${contextUrl}${userBaseURL}/roles`,
-		onNetworkStatusChange: setNetworkStatus,
-		variables: {
-			pageSize: -1,
-		},
-	});
-
 	useEffect(() => {
-		retrieveAccountRoles(accountEntryId)
-			.then((response) => response.json())
-			.then(({items}) => {
-				const accountRoleItems = items.map(({displayName, name}) => {
-					return {
-						roleKey: name,
-						roleName: displayName,
-						roleType: 'Account',
-					};
-				});
-
-				setAccountRoles(accountRoleItems);
-			});
-
 		const sectionsData = [];
 
 		if (notifications) {
@@ -88,12 +48,10 @@ const Notifications = (props) => {
 	return sections.map(({identifier}, index) => (
 		<NotificationsInfo
 			{...props}
-			accountRoles={accountRoles}
+			accountEntryId={accountEntryId}
 			identifier={identifier}
 			index={index}
 			key={`section-${identifier}`}
-			networkStatus={networkStatus}
-			resource={resource}
 			sectionsLength={sections?.length}
 			setSections={setSections}
 		/>
